requirements.txt
torch>=1.7.0
numpy
opencv-python
matplotlib
PyYAML
pillow
tqdm
seaborn
scipy
